requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
altgraph==0.17
    # via
    #   macholib
    #   pyinstaller
astroid==2.4.2
    # via pylint
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
deprecated==1.2.10
    # via pygithub
ecdsa==0.16.1
    # via esptool
esptool==2.8
    # via -r requirements.in
idna==2.10
    # via requests
isort==5.7.0
    # via pylint
lazy-object-proxy==1.4.3
    # via astroid
loguru==0.5.3
    # via -r requirements.in
macholib==1.14
    # via pyinstaller
mccabe==0.6.1
    # via pylint
numpy==1.19.5
    # via wxpython
pillow==8.1.0
    # via wxpython
pyaes==1.6.1
    # via esptool
pygithub==1.54.1
    # via -r requirements.in
pyinstaller-hooks-contrib==2020.11
    # via pyinstaller
pyinstaller==4.1
    # via -r requirements.in
pyjwt==1.7.1
    # via pygithub
pylint==2.6.0
    # via -r requirements.in
pyserial==3.5
    # via esptool
requests==2.25.1
    # via
    #   -r requirements.in
    #   pygithub
six==1.15.0
    # via
    #   astroid
    #   ecdsa
    #   wxpython
toml==0.10.2
    # via pylint
typed-ast==1.4.2
    # via astroid
urllib3==1.26.2
    # via requests
wheel==0.36.2
    # via -r requirements.in
wrapt==1.12.1
    # via
    #   astroid
    #   deprecated
wxpython==4.1.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
